Update internal_standards_evaluation.Rmd
Calculate the neutral mass according to the column "formula" (with the package "Rdisop") and the corresponding adduct (with the package "CompoundDb") indicated in the column "POS" or "NEG" (depending on which ionization mode we are working).
For individual plots, only plot the relative intensities.
Add boxplots with all samples and compounds together.